Move BuildExecutionTimer out of baseServices into core, and into a better package.
